firmware/python/bitkey/fwa/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bitkey-fwa"
version = "1.2.1"
description = "Bitkey firmware analysis"
readme = "README.md"
authors = [{ name = "HWSEC" }]
dependencies = [
  "click==8.1.7",
  "pyelftools==0.32",
  "pytz==2023.3.post1",
  "pycryptodome==3.20.0",
  "PyYAML==6.0.1",
]

[project.scripts]
bitkey-fwa = "bitkey_fwa.__main__:cli"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
include = ["bitkey_fwa"]
